requirements.txt
fastapi 
uvicorn
langchain
langchain-openai
langchain-core
openai
pydantic
python-dotenv
pytest
pytest-cov